refactor: move island_systems and island_components to demo directory
These files are island-demo-specific and don't belong in the shared engine library.
