+++ src/engine/app_base.cpp
@@ -1,7 +1,7 @@
 #include "gseurat/engine/app_base.hpp"
 #include "gseurat/engine/gaussian_cloud.hpp"
 #include "gseurat/engine/gs_vfx.hpp"
-#include "gseurat/engine/island_components.hpp"
+#include "gseurat/demo/island_components.hpp"
 
 #define GLFW_INCLUDE_VULKAN
 #include <GLFW/glfw3.h>
